The host-process guards were 2,020 lines in one file
Split the way the code they cover is: ports, logs, manifest, setups,
dependencies, shutdown, process groups, ledger, stop, startup, restart,
and the guard that keeps unix-only tests off the Windows job.

Three of the twelve hold nothing but `#[cfg(unix)]` tests, which leaves
the module empty on Windows and its `use super::*` with nothing to
import -- an error under `-D warnings`, and only on the platform this
machine does not build by default. `make desktop-check-windows` found
all three.
