+++ desktop/scripts/file-size-baseline.json
@@ -10,7 +10,6 @@
   "desktop/agent-host/tests/real_harness_e2e.rs": 1202,
   "desktop/agent-host/tests/support/mod.rs": 1417,
   "desktop/locald/src/agent_host.rs": 1564,
-  "desktop/locald/src/host_process.rs": 4688,
   "desktop/locald/src/managed_runtime.rs": 2239,
   "desktop/locald/src/native_host_pack.rs": 1968,
   "desktop/locald/src/operator_config.rs": 2423,
